#100: move some parts of external libraries to the n4js.ui
