+++ plugins/org.eclipse.n4js.ui/META-INF/MANIFEST.MF
@@ -62,6 +62,7 @@ Export-Package: org.eclipse.n4js.ui,
  org.eclipse.n4js.ui.editor.syntaxcoloring,
  org.eclipse.n4js.ui.export,
  org.eclipse.n4js.ui.export.nfar,
+ org.eclipse.n4js.ui.external,
  org.eclipse.n4js.ui.formatting2,
  org.eclipse.n4js.ui.handler,
  org.eclipse.n4js.ui.internal,
