+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/external/ExternalLibraryBuildJobProvider.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.external;
+package org.eclipse.n4js.ui.external;
 
 import static com.google.common.base.Preconditions.checkNotNull;
