+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/internal/ContributingModule.java
@@ -14,6 +14,54 @@ import static com.google.inject.Scopes.SINGLETON;
 
 import org.eclipse.core.resources.IWorkspace;
 import org.eclipse.core.resources.IWorkspaceRoot;
+import org.eclipse.n4js.binaries.BinariesPreferenceStore;
+import org.eclipse.n4js.binaries.BinariesProvider;
+import org.eclipse.n4js.binaries.BinariesValidator;
+import org.eclipse.n4js.binaries.BinaryCommandFactory;
+import org.eclipse.n4js.binaries.OsgiBinariesPreferenceStore;
+import org.eclipse.n4js.binaries.nodejs.NodeBinaryLocatorHelper;
+import org.eclipse.n4js.binaries.nodejs.NodeJsBinary;
+import org.eclipse.n4js.binaries.nodejs.NodeProcessBuilder;
+import org.eclipse.n4js.binaries.nodejs.NpmBinary;
+import org.eclipse.n4js.binaries.nodejs.NpmrcBinary;
+import org.eclipse.n4js.external.EclipseTargetPlatformInstallLocationProvider;
+import org.eclipse.n4js.external.ExternalLibraryUriHelper;
+import org.eclipse.n4js.external.ExternalLibraryWorkspace;
+import org.eclipse.n4js.external.ExternalProjectCacheLoader;
+import org.eclipse.n4js.external.ExternalProjectProvider;
+import org.eclipse.n4js.external.ExternalProjectsCollector;
+import org.eclipse.n4js.external.GitCloneSupplier;
+import org.eclipse.n4js.external.RebuildWorkspaceProjectsScheduler;
+import org.eclipse.n4js.external.TargetPlatformInstallLocationProvider;
+import org.eclipse.n4js.external.TypeDefinitionGitLocationProvider;
+import org.eclipse.n4js.external.TypeDefinitionGitLocationProvider.TypeDefinitionGitLocationProviderImpl;
+import org.eclipse.n4js.internal.FileBasedExternalPackageManager;
+import org.eclipse.n4js.internal.InternalN4JSWorkspace;
+import org.eclipse.n4js.internal.N4JSModel;
+import org.eclipse.n4js.preferences.ExternalLibraryPreferenceStore;
+import org.eclipse.n4js.preferences.OsgiExternalLibraryPreferenceStore;
+import org.eclipse.n4js.projectModel.IN4JSCore;
+import org.eclipse.n4js.ui.containers.CompositeStorage2UriMapperContribution;
+import org.eclipse.n4js.ui.containers.N4JSExternalLibraryStorage2UriMapperContribution;
+import org.eclipse.n4js.ui.containers.N4JSToBeBuiltComputer;
+import org.eclipse.n4js.ui.containers.NfarStorageMapper;
+import org.eclipse.n4js.ui.external.EclipseExternalLibraryWorkspace;
+import org.eclipse.n4js.ui.external.ExternalLibraryBuildJobProvider;
+import org.eclipse.n4js.ui.external.ExternalLibraryBuilderHelper;
+import org.eclipse.n4js.ui.external.ProjectStateChangeListener;
+import org.eclipse.n4js.ui.navigator.N4JSProjectExplorerLabelProvider;
+import org.eclipse.n4js.ui.navigator.internal.N4JSProjectExplorerHelper;
+import org.eclipse.n4js.ui.projectModel.IN4JSEclipseCore;
+import org.eclipse.n4js.ui.scoping.builtin.ScopeInitializer;
+import org.eclipse.n4js.ui.workingsets.WorkingSetManagerBroker;
+import org.eclipse.n4js.ui.workingsets.WorkingSetManagerBrokerImpl;
+import org.eclipse.n4js.ui.workingsets.WorkingSetManagerModificationStrategyProvider;
+import org.eclipse.n4js.ui.workingsets.WorkingSetManualAssociationWizard;
+import org.eclipse.n4js.ui.workingsets.WorkingSetProjectNameFilterWizard;
+import org.eclipse.n4js.utils.StatusHelper;
+import org.eclipse.n4js.utils.process.OutputStreamPrinterThreadProvider;
+import org.eclipse.n4js.utils.process.OutputStreamProvider;
+import org.eclipse.n4js.utils.process.ProcessExecutor;
 import org.eclipse.xtext.builder.builderState.IBuilderState;
 import org.eclipse.xtext.builder.clustering.CurrentDescriptions;
 import org.eclipse.xtext.builder.impl.IToBeBuiltComputerContribution;
@@ -32,55 +80,6 @@ import com.google.inject.Inject;
 import com.google.inject.Module;
 import com.google.inject.Provider;
 import com.google.inject.name.Names;
-
-import org.eclipse.n4js.binaries.BinariesPreferenceStore;
-import org.eclipse.n4js.binaries.BinariesProvider;
-import org.eclipse.n4js.binaries.BinariesValidator;
-import org.eclipse.n4js.binaries.BinaryCommandFactory;
-import org.eclipse.n4js.binaries.OsgiBinariesPreferenceStore;
-import org.eclipse.n4js.binaries.nodejs.NodeBinaryLocatorHelper;
-import org.eclipse.n4js.binaries.nodejs.NodeJsBinary;
-import org.eclipse.n4js.binaries.nodejs.NodeProcessBuilder;
-import org.eclipse.n4js.binaries.nodejs.NpmBinary;
-import org.eclipse.n4js.binaries.nodejs.NpmrcBinary;
-import org.eclipse.n4js.external.EclipseExternalLibraryWorkspace;
-import org.eclipse.n4js.external.EclipseTargetPlatformInstallLocationProvider;
-import org.eclipse.n4js.external.ExternalLibraryBuildJobProvider;
-import org.eclipse.n4js.external.ExternalLibraryBuilderHelper;
-import org.eclipse.n4js.external.ExternalLibraryUriHelper;
-import org.eclipse.n4js.external.ExternalLibraryWorkspace;
-import org.eclipse.n4js.external.ExternalProjectCacheLoader;
-import org.eclipse.n4js.external.ExternalProjectProvider;
-import org.eclipse.n4js.external.ExternalProjectsCollector;
-import org.eclipse.n4js.external.GitCloneSupplier;
-import org.eclipse.n4js.external.ProjectStateChangeListener;
-import org.eclipse.n4js.external.RebuildWorkspaceProjectsScheduler;
-import org.eclipse.n4js.external.TargetPlatformInstallLocationProvider;
-import org.eclipse.n4js.external.TypeDefinitionGitLocationProvider;
-import org.eclipse.n4js.external.TypeDefinitionGitLocationProvider.TypeDefinitionGitLocationProviderImpl;
-import org.eclipse.n4js.internal.FileBasedExternalPackageManager;
-import org.eclipse.n4js.internal.InternalN4JSWorkspace;
-import org.eclipse.n4js.internal.N4JSModel;
-import org.eclipse.n4js.preferences.ExternalLibraryPreferenceStore;
-import org.eclipse.n4js.preferences.OsgiExternalLibraryPreferenceStore;
-import org.eclipse.n4js.projectModel.IN4JSCore;
-import org.eclipse.n4js.ui.containers.CompositeStorage2UriMapperContribution;
-import org.eclipse.n4js.ui.containers.N4JSExternalLibraryStorage2UriMapperContribution;
-import org.eclipse.n4js.ui.containers.N4JSToBeBuiltComputer;
-import org.eclipse.n4js.ui.containers.NfarStorageMapper;
-import org.eclipse.n4js.ui.navigator.N4JSProjectExplorerLabelProvider;
-import org.eclipse.n4js.ui.navigator.internal.N4JSProjectExplorerHelper;
-import org.eclipse.n4js.ui.projectModel.IN4JSEclipseCore;
-import org.eclipse.n4js.ui.scoping.builtin.ScopeInitializer;
-import org.eclipse.n4js.ui.workingsets.WorkingSetManagerBroker;
-import org.eclipse.n4js.ui.workingsets.WorkingSetManagerBrokerImpl;
-import org.eclipse.n4js.ui.workingsets.WorkingSetManagerModificationStrategyProvider;
-import org.eclipse.n4js.ui.workingsets.WorkingSetManualAssociationWizard;
-import org.eclipse.n4js.ui.workingsets.WorkingSetProjectNameFilterWizard;
-import org.eclipse.n4js.utils.StatusHelper;
-import org.eclipse.n4js.utils.process.OutputStreamPrinterThreadProvider;
-import org.eclipse.n4js.utils.process.OutputStreamProvider;
-import org.eclipse.n4js.utils.process.ProcessExecutor;
 
 /**
  */
