+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/preferences/external/ExternalLibraryPreferencePage.java
@@ -14,6 +14,8 @@ import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.base.Strings.isNullOrEmpty;
 import static com.google.common.collect.FluentIterable.from;
 import static com.google.common.primitives.Ints.asList;
+import static java.util.Collections.singletonList;
+import static org.eclipse.jface.layout.GridDataFactory.fillDefaults;
 import static org.eclipse.n4js.external.libraries.ExternalLibrariesActivator.EXTERNAL_LIBRARIES_SUPPLIER;
 import static org.eclipse.n4js.external.libraries.ExternalLibrariesActivator.N4_NPM_FOLDER_SUPPLIER;
 import static org.eclipse.n4js.external.libraries.ExternalLibrariesActivator.repairNpmFolderState;
@@ -21,8 +23,6 @@ import static org.eclipse.n4js.external.libraries.TargetPlatformModel.TP_FILTER_
 import static org.eclipse.n4js.ui.preferences.external.ButtonFactoryUtil.createDisabledPushButton;
 import static org.eclipse.n4js.ui.preferences.external.ButtonFactoryUtil.createEnabledPushButton;
 import static org.eclipse.n4js.ui.utils.DelegatingSelectionAdapter.createSelectionListener;
-import static java.util.Collections.singletonList;
-import static org.eclipse.jface.layout.GridDataFactory.fillDefaults;
 import static org.eclipse.swt.SWT.END;
 import static org.eclipse.swt.SWT.FILL;
 import static org.eclipse.swt.SWT.OPEN;
@@ -55,6 +55,27 @@ import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.viewers.SelectionChangedEvent;
 import org.eclipse.jface.viewers.TreePath;
 import org.eclipse.jface.viewers.TreeViewer;
+import org.eclipse.n4js.binaries.IllegalBinaryStateException;
+import org.eclipse.n4js.external.ExternalLibraryWorkspace;
+import org.eclipse.n4js.external.GitCloneSupplier;
+import org.eclipse.n4js.external.NpmManager;
+import org.eclipse.n4js.external.TargetPlatformInstallLocationProvider;
+import org.eclipse.n4js.external.libraries.TargetPlatformModel;
+import org.eclipse.n4js.external.version.VersionConstraintFormatUtil;
+import org.eclipse.n4js.n4mf.DeclaredVersion;
+import org.eclipse.n4js.n4mf.ProjectDescription;
+import org.eclipse.n4js.n4mf.utils.parsing.ManifestValuesParsingUtil;
+import org.eclipse.n4js.n4mf.utils.parsing.ParserResults;
+import org.eclipse.n4js.preferences.ExternalLibraryPreferenceStore;
+import org.eclipse.n4js.projectModel.IN4JSProject;
+import org.eclipse.n4js.ui.external.ExternalLibrariesReloadHelper;
+import org.eclipse.n4js.ui.utils.InputComposedValidator;
+import org.eclipse.n4js.ui.utils.InputFunctionalValidator;
+import org.eclipse.n4js.ui.utils.UIUtils;
+import org.eclipse.n4js.ui.viewer.TreeViewerBuilder;
+import org.eclipse.n4js.utils.StatusHelper;
+import org.eclipse.n4js.utils.collections.Arrays2;
+import org.eclipse.n4js.utils.io.FileDeleter;
 import org.eclipse.swt.events.SelectionEvent;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
@@ -69,28 +90,6 @@ import org.eclipse.ui.IWorkbenchPreferencePage;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
-
-import org.eclipse.n4js.binaries.IllegalBinaryStateException;
-import org.eclipse.n4js.external.ExternalLibrariesReloadHelper;
-import org.eclipse.n4js.external.ExternalLibraryWorkspace;
-import org.eclipse.n4js.external.GitCloneSupplier;
-import org.eclipse.n4js.external.NpmManager;
-import org.eclipse.n4js.external.TargetPlatformInstallLocationProvider;
-import org.eclipse.n4js.external.libraries.TargetPlatformModel;
-import org.eclipse.n4js.external.version.VersionConstraintFormatUtil;
-import org.eclipse.n4js.n4mf.DeclaredVersion;
-import org.eclipse.n4js.n4mf.ProjectDescription;
-import org.eclipse.n4js.n4mf.utils.parsing.ManifestValuesParsingUtil;
-import org.eclipse.n4js.n4mf.utils.parsing.ParserResults;
-import org.eclipse.n4js.preferences.ExternalLibraryPreferenceStore;
-import org.eclipse.n4js.projectModel.IN4JSProject;
-import org.eclipse.n4js.ui.utils.InputComposedValidator;
-import org.eclipse.n4js.ui.utils.InputFunctionalValidator;
-import org.eclipse.n4js.ui.utils.UIUtils;
-import org.eclipse.n4js.ui.viewer.TreeViewerBuilder;
-import org.eclipse.n4js.utils.StatusHelper;
-import org.eclipse.n4js.utils.collections.Arrays2;
-import org.eclipse.n4js.utils.io.FileDeleter;
 
 /**
  * Preference page for managing external libraries.
